requirements.txt
# Azure Functions
azure-functions>=1.12.0

# HTTP requests
requests>=2.28.0

# Azure Storage Queue
azure-storage-queue>=12.5.0

# Azure Identity
azure-identity>=1.12.0

# Utilities
python-dotenv>=1.0.0
